% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2px.R
\name{data2px}
\alias{data2px}
\title{Trasforma un cubo di dati da data frame a oggetto di classe 'px'}
\usage{
data2px(.data, id.vars, measure.vars, variable.name = "misura")
}
\arguments{
\item{.data}{cubo di dati in forma di data frame}

\item{id.vars}{vettore character con i nomi delle variabili categoriali del cubo di dati}

\item{measure.vars}{vettore character con i nomi delle colonne che contengono le statistiche del cubo}

\item{variable.name}{nome da dare alla colonna che conterrà il nome della misura (default = "misura")}
}
\description{
Trasforma un cubo di dati da data frame a oggetto di classe 'px'
}
\examples{
data(c09C)

str(c09C)
head(c09C)

# Trasforma il data frame c09C in oggetto di classe 'px'
c09C_px <- data2px(c09C, 
                   id.vars = c("anno", "posizione", "sesso", "statuto"),
                   measure.vars = c("p10", "p25", "p50", "p75", "p90"))
                   
class(c09C_px)

}

