% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_tidy2px.R
\docType{package}
\name{tidy2px}
\alias{tidy2px}
\alias{tidy2px-package}
\title{Tools to estimate and manage empirical distributions}
\description{
Tools to estimate and manage empirical distributions,
        which should work with survey data. It's possible to have the
        same functionalities in different ways (for example based on
        plyr and dplyr).
}

