% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_tidy2px.R
\docType{package}
\name{tidy2px}
\alias{tidy2px}
\alias{tidy2px-package}
\title{Trasforma un Cubo di Dati da Data Frame a File '.px'}
\description{
Trasforma un cubo di dati da data frame a file '.px'. Prima di
essere trasformato in oggetto e file '.px', il cubo in forma di data frame viene
trasformato in array.
}
